# educa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Educa LRS REST API Backend",
        "description": "Educa LRS API Backend",
        "contact": {
            "email": "support@digitallearning.gmbh"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/objects": {
            "get": {
                "tags": [
                    "Objects"
                ],
                "summary": "List all objects",
                "operationId": "d7a2d9291e5f8b61329bb8a474a89920",
                "responses": {
                    "200": {
                        "description": "List of objects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "definition": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Module 1"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "module"
                                                    },
                                                    "iri": {
                                                        "type": "string",
                                                        "example": "http://example.com/modules/1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Objects"
                ],
                "summary": "Create a new object",
                "operationId": "1cde3c97246d6bb8dd55334720b9a7e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "iri"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Module 1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "module"
                                    },
                                    "iri": {
                                        "type": "string",
                                        "example": "http://example.com/modules/1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Object created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "definition": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Module 1"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "module"
                                                },
                                                "iri": {
                                                    "type": "string",
                                                    "example": "http://example.com/modules/1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/objects/{id}": {
            "get": {
                "tags": [
                    "Objects"
                ],
                "summary": "Retrieve an object by ID",
                "operationId": "00769e257b8ac6644390da518e85c420",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the object to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "definition": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Module 1"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "module"
                                                },
                                                "iri": {
                                                    "type": "string",
                                                    "example": "http://example.com/modules/1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Objects"
                ],
                "summary": "Update an object",
                "operationId": "3b03ef8c1b993e9951d9c40022443da3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the object to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Module"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "updated-module"
                                    },
                                    "iri": {
                                        "type": "string",
                                        "example": "http://example.com/modules/updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Object updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "definition": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Updated Module"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "updated-module"
                                                },
                                                "iri": {
                                                    "type": "string",
                                                    "example": "http://example.com/modules/updated"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Objects"
                ],
                "summary": "Delete an object",
                "operationId": "9edaeea5b896c536e48582c95b8af81a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the object to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}